Attributes

Setting individual attrributes

SimpleDrawingObjects.set_fillalpha! — Function

set_fillalpha!(o::FilledObject, α=1.0)

Set the transparency of o to α with 1.0 being completely opaque and 0.0 being fully transparent.

SimpleDrawingObjects.set_fillcolor! — Function

set_fillcolor!(o::FilledObject, col=:white)

Change the fill color (for FilledObjects)

SimpleDrawingObjects.set_linecolor! — Function

set_linecolor!(o::SimpleDrawingObject, c=:black)

Assign color c to the lines in this object.

SimpleDrawingObjects.set_linewidth! — Function

set_linewidth!(o::SimpleDrawingObject, w=1)

Assign line width w to this object.

SimpleDrawingObjects.set_linestyle! — Function

set_linestyle!(o::SimpleDrawingObject, s=:solid)

Assign s to be the line style of this object.

SimpleDrawingObjects.set_pointcolor! — Function

set_pointcolor!(p::Point, col = :black)

Set the color to draw a Point. The functions set_fillcolor! and set_linecolor! don't apply.

SimpleDrawingObjects.set_pointsize! — Function

set_pointsize!(p::Point, sz = 2)

Set the size at which a Point is drawn.

Resetting

SimpleDrawingObjects.reset_attributes! — Function

reset_attributes!(o::SimpleDrawingObject)

Reset o's attributes to factory defaults.

Custom control

SimpleDrawingObjects.set_attribute! — Function

set_attribute!(o::SimpleDrawingObject, attr::Symbol, val)

Assign the value val to the object's attribute attr.

SimpleDrawingObjects.get_attributes — Function

get_attributes(o::SimpleDrawingObject)::Dict

Give access to the attributes dictionary for o.

Line Segments

SimpleDrawingObjects.Segment — Type

Segment(a::Number, b::Number)
Segment(x::Real, y::Real, xx::Real, yy::Real)

Create a segment from point a to point b where a and b are complex numbers.

In the second form, create a line segment from x + y*im to xx + yy*im.

SimpleDrawingObjects.Arrow — Function

Arrow(a::Number, b::Number)::Segment
Arrow(x::Real, y::Real, xx::Real, yy::Real)::Segment

Create a line segment with the added attribute of having an arrow head at b (at (xx,yy)).

Polygons

SimpleDrawingObjects.Polygon — Type

Polygon(plist::Vector{T}) where {T<:Number}
Polygon(zs...)

Create a polygon from a list of complex numbers. Polygon([0, 2im, 4]) can also be created using Polygon(0, 2im, 4).

SimpleDrawingObjects.FilledPolygon — Type

FilledPolygon(plist::Vector{T}) where {T<:Number}
FilledPolygon(zs...)

Create a filled-in polygon from a list of complex numbers. FilledPolygon([0, 2im, 4]) can also be created using FilledPolygon(0, 2im, 4).

SimpleDrawingObjects.Rectangle — Function

Rectangle(a::Number, b::Number)::Polygon
Rectangle(x::Real, y::Real, xx::Real, yy::Real)::Polygon

Create a rectangle with opposite corners a and b [or (x,y) and (xx,yy)].

SimpleDrawingObjects.FilledRectangle — Function

FilledRectangle(a::Number, b::Number)::FilledPolygon
FilledRectangle(x::Real, y::Real, xx::Real, yy::Real)::FilledPolygon

Create a rectangle with opposite corners a and b [or (x,y) and (xx,yy)].

Circular Shapes

SimpleDrawingObjects.Arc — Type

Arc(z::Number, r::Real, t1::Real, t2::Real, t3::Real)
Arc(x::Real, y::Real, r::Real, t1::Real, t2::Real, t3::Real)

Create an arc centered at z [at (x,y)] with radius r starting at angle t1, continuing through angle t2, and ending at angle t3.

SimpleDrawingObjects.Circle — Type

Circle(z::Number, r::Real)
Circle(x::Real, y::Real, r::Real)

Create a circle centered at z [at (x,y)] with radius r.

SimpleDrawingObjects.FilledCircle — Type

Disk(z::Number, r::Real)
Disk(x::Real, y::Real, r::Real)

Create a disk, i.e., a filled circle with default fill color set to :white.

See: set_fillcolor!.

SimpleDrawingObjects.Ellipse — Type

Ellipse(x::Real, y::Real, rx::Real, ry::Real)
Ellipse(x::Real, y::Real, rx::Real, ry::Real)

Create an ellipse centered at z [at (x,y)] with radii rx and ry.

SimpleDrawingObjects.FilledEllipse — Type

FilledEllipse(x::Real, y::Real, rx::Real, ry::Real)
FilledEllipse(x::Real, y::Real, rx::Real, ry::Real)

Create a filled ellipse centered at z [at (x,y)] with radii rx and ry.

Spline Curves

SimpleDrawingObjects.OpenCurve — Type

OpenCurve(plist::Vector{T}) where {T<:Number}
OpenCurve(zs...)

Draw an open curve using the given points using a cubic spline.

SimpleDrawingObjects.ClosedCurve — Type

ClosedCurve(plist::Vector{T}) where {T<:Number}
ClosedCurve(zs...)

Draw a closed curve using the given points using a cubic spline.

SimpleDrawingObjects.FilledClosedCurve — Type

FilledClosedCurve(plist::Vector{T}) where {T<:Number}
FilledClosedCurve(zs...)

Draw a closed curve using the given points using a cubic spline whose interior is colored.

Points

SimpleDrawingObjects.Point — Type

Point(a::Number)
Point(x::Real, y::Real)

Create a point at location a [or (x,y)].

Types

SimpleDrawingObjects.SimpleDrawingObject — Type

SimpleDrawingObject

Parent abstract type for all geometric objects defined in SimpleDrawingObjects.

SimpleDrawingObjects.FilledObject — Type

FilledObject

Subtype of SimpleDrawingObject for geometric objects that can be filled with color.

Index

SimpleDrawingObjects.Arc

SimpleDrawingObjects.Circle

SimpleDrawingObjects.ClosedCurve

SimpleDrawingObjects.Ellipse

SimpleDrawingObjects.FilledCircle

SimpleDrawingObjects.FilledClosedCurve

SimpleDrawingObjects.FilledEllipse

SimpleDrawingObjects.FilledObject

SimpleDrawingObjects.FilledPolygon

SimpleDrawingObjects.OpenCurve

SimpleDrawingObjects.Point

SimpleDrawingObjects.Polygon

SimpleDrawingObjects.Segment

SimpleDrawingObjects.SimpleDrawingObject

SimpleDrawingObjects.Arrow

SimpleDrawingObjects.FilledRectangle

SimpleDrawingObjects.Rectangle

SimpleDrawingObjects.get_attributes

SimpleDrawingObjects.reset_attributes!

SimpleDrawingObjects.set_attribute!

SimpleDrawingObjects.set_fillalpha!

SimpleDrawingObjects.set_fillcolor!

SimpleDrawingObjects.set_linecolor!

SimpleDrawingObjects.set_linestyle!

SimpleDrawingObjects.set_linewidth!

SimpleDrawingObjects.set_pointcolor!

SimpleDrawingObjects.set_pointsize!